Add blacklist option to config to ignore files
You can now setup a blacklist in your config to ignore a set of files.
Currently, the blacklist supports file paths or directory paths. For
example, let's say you have the following directory structure:

```bash
└── project
    ├── .eslintrc
    ├── index.js
    ├── jquery.js
    └── lib
        └── underscore.js
```

And the contents of your `.eslintrc`:

```json
{
  ...

  "blacklist": ["jquery.js", lib]
}

When you run `eslint`, the files and folders in the blacklist will be ignored:

```bash
└── project
    ├── .eslintrc
    ├── index.js [LINTED]
    ├── jquery.js [IGNORED]
    └── lib
        └── underscore.js [IGNORED]
```